requirements.txt
asgiref==3.10.0
boto3==1.41.1
botocore==1.41.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
pillow==12.0.0
psycopg2==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
